+++ src/app/app.module.ts
@@ -77,6 +77,7 @@ import { UserManagementComponent } from './user-management/user-management.compo
 import { RefDatumEditComponent } from './ref-datum-edit/ref-datum-edit.component';
 import { RoleService } from './services/role.service';
 import { AgencyService } from './services/agency.service';
+import { FileEditComponent } from './file-edit/file-edit.component';
 import { HwmEditComponent } from './hwm-edit/hwm-edit.component';
 import { PeakEditComponent } from './peak-edit/peak-edit.component';
 import { SessionEventComponent } from './session-event/session-event.component';
@@ -110,6 +111,7 @@ import { SessionEventComponent } from './session-event/session-event.component';
         HwmEditComponent,
         SensorEditComponent,
         PeakEditComponent,
+        FileEditComponent,
         SessionEventComponent,
     ],
     imports: [